BUGFIX 1: The code has been adapted to the cprover framework update.
BUGFIX 2:
A bug has been found in the upgraded version of cprover framework. It seems that the instruction
t3 = t1 - t2
not only assigns to t3 the difference between t1 and t2, but it assigns this difference to t1 too!
This cprover bug generated some problems in dependency_checker code, but these problems has been fixed by modifying the instructions to calculate time performances in order to avoid this problem. The new instructions are such that even fixing the cprover code they will still work properly.

IMPROVEMENT:
(BEFORE)
In many cases the find_var_deps() function was not efficient and sometimes its running time for big benchmarks was even more than solving time, interpolation time and conversion time. In the past an attempt to implement a basic union/find function has been made to make it more efficient, but the complexity had been reduced just for few cases. Moreover that union/find approach was not correct and this has discovered by using a counterexample.
(NOW)
A popular and efficient union/find structure has been used from the open library boost.org, now the approach is correct and it is even very efficient, now the complexity of find_var_deps() is probabilistically linear in terms of the size of the sequence of variables which appear in the SSA steps sequence. In tests, the cases in which find_var_deps() took, say 100 seconds, now it takes just 2.5 seconds.


git-svn-id: https://scm.ti-edu.ch/reposvn/verify_cprover@340 533bbb3b-ce99-0410-9130-e5b668893301